Factorisation hr
